requirements.txt
attrs==21.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.1
cycler==0.10.0
docutils==0.18.1
idna==3.3
importlib-metadata==4.11.2
impyute==0.0.8
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.0.1
joblib==1.1.0
keyring==23.5.0
kiwisolver==1.3.2
lightgbm==3.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
miceforest==5.3.0
missingno==0.5.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
patsy==0.5.2
Pillow==8.3.2
pkginfo==1.8.2
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.21
Pygments==2.11.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
readme-renderer==32.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
SecretStorage==3.3.1
six==1.16.0
sklearn==0.0
statsmodels==0.13.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.62.3
twine==3.8.0
urllib3==1.26.8
webencodings==0.5.1
xgboost==1.6.1
yattag==1.14.0
zipp==3.7.0
